begin to add new page for decks

+++ TestTcgScanner/TestTcgScanner/Models/MtgCard.cs
@@ -12,7 +12,7 @@ namespace TestTcgScanner
         public string? Id { get; set; }
         public List<string>? Names { get; set; }
         public string ManaCost { get; set; } = "";
-        public int? Cmc {  get; set; }
+        public double? Cmc {  get; set; }
         public List<string>? Colors {  get; set; }
         public List<string>? ColorIdentity { get; set; }
         public string Type { get; set; } = "";
